+++ NEWS.md
@@ -1,4 +1,7 @@
 # isopleuros 1.0.0.9000
+## New classes and methods
+* Add `ternary_labels()` to draw non-overlapping text labels.
+
 ## Internals
 * Use **tinytest** and **tinysnapshot** instead of **testthat** and **vdiffr**.